replace GuiRendererHelper with QuadBuilder
update all rendering to use QuadBatchManager + QuadBuilder
