+++ library/rendering/item_rendering/src/main/java/org/quiltmc/qsl/rendering/item/api/client/SolidColorItemBarRenderer.java
@@ -16,9 +16,6 @@
 
 package org.quiltmc.qsl.rendering.item.api.client;
 
-import com.mojang.blaze3d.systems.RenderSystem;
-import com.mojang.blaze3d.vertex.BufferBuilder;
-import com.mojang.blaze3d.vertex.Tessellator;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 
@@ -45,16 +42,9 @@ public abstract class SolidColorItemBarRenderer implements ItemBarRenderer {
 
 	@Override
 	public void renderItemBar(MatrixStack matrices, QuadBatchManager quadBatchManager, TextRenderer textRenderer, float zOffset, ItemStack stack) {
-		RenderSystem.disableDepthTest();
-		RenderSystem.disableTexture();
-		RenderSystem.disableBlend();
-		Tessellator tessellator = Tessellator.getInstance();
-		BufferBuilder buffer = tessellator.getBufferBuilder();
-		GuiRendererHelper.renderQuad(matrices, buffer, 2, 0, 13, 2, getItemBarBackground(stack));
-		GuiRendererHelper.renderQuad(matrices, buffer, 2, 0, getItemBarStep(stack), 1, getItemBarForeground(stack));
-		RenderSystem.enableBlend();
-		RenderSystem.enableTexture();
-		RenderSystem.enableDepthTest();
+		var buffer = quadBatchManager.beginQuads();
+		QuadBuilder.add(matrices, buffer, 2, 0, 13, 2, getItemBarBackground(stack));
+		QuadBuilder.add(matrices, buffer, 2, 0, getItemBarStep(stack), 1, getItemBarForeground(stack));
 	}
 
 	/**
